Initial code commit
This is the initial code commit. As of now, the application implements the [Code scanner library](https://github.com/yuriy-budiyev/code-scanner) as well as the [RTK library](https://github.com/RastislavKish/RtkKotlinAndroid)
It should be capable of requesting camera permission, scanning for barcodes and beeping when one is found, announcing its value.