Add logical background for MainActivity
Added the central RScan class, responsible for program logics. The UI of the main activity is now fully functional, capable of either scanning barcodes of one of the selectable types, or storing scanning results in a recycler view, where it will be later possible to register them.
The list is cleared automatically when user leaves the discovery mode.
BarcodeInfo now also defines the equals method, so it is possible to compare barcodes simply by the == sign, or work with them in collections with their search mechanisms. A match occurs in case, that compared barcodes are of identical type and value.