Add barcode identification activity
Added the barcode identification activity. It currently consists of 4 guidelines (two horizontal and two vertical), splitting the screen to upper TextView containing information about the barcode being identified, a RecyclerView for suggested barcode descriptions and a lower bar, which contains an EditText for entering the final description, a word counter for using just specified number of first words from the entered description and finally a Save button, which should Save it and return to the main activity.