Add background logic for BarcodeIdentificationActivity
Added logic for BarcodeIdentificationActivity, including barcode lookups on DuckDuckGo, selecting a number of words from the input EditText and sending the result back to the MainActivity.
In MainActivity, clicking on a scanning result in scanningResultsRecyclerView now opens the identification activity with the clicked barcode, the result is not processed yet, though.
A bug in BScanner causing all recognized barcodes other than EAN-13 being ignored was fixed, so the Discovery scanning mode should work properly now.