Remove scanning modes and result, redesign MainActivity
Scanning modes were removed, as they does not seem practical in the normal usage.
Along with this change, with the discovery mode becoming the default, the scanning result TextView become also useless, as well as the whole upper bar.
MainActivity thus got a major redesign, taking away the upper bar and creating a newone in the bottom.
It contains a button to Clear the list (which is now the only place for storing scanning results, with new popping on the top) and a flashlight usage ToggleButton, allowing the user to dynamically turn the torch on and off.