Add barcodes caching
BarcodeCache was added, allowing barcodes to be described by the user. Each described barcode is saved to SharedPreferences in the csv format, all related logic is in the BarcodeCache class.
UI now works with this feature, including the Save button in BarcodeIdentificationActivity as well as the recognition itself.

From minor changes, BarcodeInfo class was separated to its own file due to its wide usage in the application.
The code was also cleared a bit, including deleting unused imports, libraries or the companion object of PermissionsRequester.