Add Fastlane metadata
Metadata were added in Fastlane format in order to cooperate better with F-Droid according to the [official reference.](https://f-droid.org/en/docs/All_About_Descriptions_Graphics_and_Screenshots/)

Currently, the structure contains:

* Title, short description and full description in English and Slovak.
* An icon image and three screenshots from the app showing a scanned code, the identification activity with that code and finally, a recognized product.
* Changelogs in English and Slovak, currently for versions 1.0 and 1.1.

readme was edited to incorporate the license of the icon, the whole Assets section got a new design with a fancy table showing everything necessary. A license header was added to its end in order to make the repository licensing independend on GitHub.

The headers in the source code contained one trailing space, which was removed.

On BarcodeIdentificationActivity, the word selector was slightly edited to center the count (it was aligned to the left previously, what looked weird).