Fix crashes caused by portrait phone orientation on app startup
A request for landscape orientation in onCreate method of an activity destroys the activity in case its orientation is different from the requestedone, a completely new activity is started with the requested orientation.
This was causing problems, since all the initialization of the original activity is performed as usual, it caused initialization of the camera on portrait MainActivity in this case.
Which was gone before the camera's lifecycle was registered, resulting in a crash.

The activity's orientation is now checked, both onCreate and onDestroy finish immediately if portraitone is detected, avoiding any initialization or deinitialization.