Allow repeated detection of describable barcodes
Allow barcodes that can be described (either via the barcode cache or the automatic description of EAN-13 / UPC-A) to be detected repeatedly, popping to the top of the results list. This should make it possible to identify multiple products of the same type without the need to clear the list between detections. Barcodes that can't be described don't get re-scanned, since the user needs to assign them a meaningful information first.